auth, authorization, movies crud frontend bugs fix

+++ app/api/src/movies/movies.service.ts
@@ -1,28 +1,56 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { MoviesDto } from './dtos/movies.dto';
 import { Movie } from '@prisma/client';
+import { SearchParams, TMovie } from './types';
 
 @Injectable()
 export class MoviesService {
   constructor(private prisma: PrismaService) {}
 
-  createMovie(data: MoviesDto): Promise<Movie> {
+  createMovie(data: TMovie, userId: number): Promise<Movie> {
+    if (data.userId !== userId) throw new ForbiddenException('Access Denied');
+
     return this.prisma.movie.create({ data });
-    // console.log(test);
   }
 
-  getMovies() {
-    return this.prisma.movie.findMany();
+  async getMovies(params: any) {
+    if (params.order === 'desc' || params.order === 'asc') {
+      return await this.prisma.movie.findMany({
+        orderBy: {
+          title: params.order,
+        },
+      });
+    } else {
+      return await this.prisma.movie.findMany();
+    }
   }
 
-  editMovie(id: number) {
-    return this.prisma.movie.findUnique({
+  async getMovie(id: number) {
+    return await this.prisma.movie.findUnique({
       where: { id },
     });
   }
 
-  updateMovie(id: number, data: MoviesDto): Promise<Movie> {
+  async editMovie(id: number, userId: number): Promise<Movie> {
+    const movie = await this.getMovie(id);
+
+    if (!movie) {
+      throw new NotFoundException('Not Found');
+    } else if (movie.userId !== userId) {
+      throw new ForbiddenException('Access Denied');
+    } else {
+      return movie;
+    }
+  }
+
+  async updateMovie(id: number, data: TMovie, userId: number): Promise<Movie> {
+    if (!userId || data.userId !== userId)
+      throw new ForbiddenException('Access Denied');
+
     return this.prisma.movie.update({
       where: {
         id,
@@ -31,9 +59,24 @@ export class MoviesService {
     });
   }
 
-  deleteMovie(id: number) {
+  async deleteMovie(id: number, userId: number): Promise<Movie> {
+    const movie = await this.getMovie(id);
+    if (!userId || movie.userId !== userId)
+      throw new ForbiddenException('Access Denied');
+
     return this.prisma.movie.delete({
       where: { id },
     });
+  }
+
+  async searchMovies(param: SearchParams) {
+    const movies = await this.prisma.movie.findMany({
+      where: {
+        title: {
+          search: param.title.split(' ').join(' & '),
+        },
+      },
+    });
+    return movies;
   }
 }
